pyproject.toml
[project]
name = "langsight"
version = "0.1.0"
description = "Open-source MCP observability and security platform"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13",
    "pydantic>=2",
    "pydantic-settings>=2",
    "structlog>=24",
    "mcp>=1.0",
    "anyio>=4",
    "pyyaml>=6",
    "httpx>=0.27",
]

[project.scripts]
langsight = "langsight.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langsight"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "ruff>=0.5",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=langsight --cov-report=term-missing --cov-fail-under=80"
markers = [
    "unit: fast tests, no external deps",
    "integration: requires docker compose up (postgres-mcp + s3-mcp)",
    "e2e: full CLI flow tests, requires full stack",
]

[tool.coverage.run]
source = ["src/langsight"]
omit = ["*/cli/*"]  # CLI tested via e2e, not unit coverage

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]
ignore = ["B008"]  # do not perform function calls in default args (Pydantic uses this)

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/langsight"]
